Add xtask to rename a binary produced by `cargo build` according to netdata requirements
Plugin binary name is required to end with ".plugin" suffix. And as it
is impossible to have a crate or produce a binary containing a dot
symbol using just cargo itself, it became necessary to wrap `cargo
build` procedure into `xtask` to rename resulting binary after the build
is complete.